requirements.txt
json5==0.12.0
lupa==2.4
pygame-ce==2.5.5
pygame_gui==0.6.14
python-i18n==0.3.9
-e git+ssh://git@github.com/97cweb/SimAntics.git@2f22b353b8671a1c29183c1fc740f268815dc7d0#egg=simantics_common&subdirectory=simantics_common
